Profile UI

+++ Controller/Runtime/Converters/ConvertTimedInfo.cs
@@ -10,10 +10,10 @@ namespace Soul.Controller.Runtime.Converters
     {
         public T data;
         public TV ratio;
-        [Range(0, 100)]public float xp = 1;
-        
+        [Range(0, 100)] public int xp = 1;
+
         public UnityTimeSpan timeRequired;
         public UnityTimeSpan RequiredTime => timeRequired;
-        public float GetXpFrom(float amount) => xp * amount;
+        public int GetXpFrom(int amount) => xp * amount;
     }
 }
